requirements.txt
Flask==3.0.0
flask_cors==4.0.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.26.0
gunicorn==21.2.0